Change the def of transfer and add iff_refl rule
The definition of transfer is changed from apply modulo to refine (modulo _)
to avoid strange error messages such as:
    Unable to satisfy the following constraints:
    In environment:
    a : int

    ?class : "Related impl (a < a) False"
Now, the tactic will leave unsolved constraints instead of failing.

A low priority iff_reflexivity rule is introduced to allow the transfer of propositions
such as False. The low priority means it will be used in last resort and does not
hinder the use of the transfer tactic.